[FRONTEND] Move Basket.js to other folder.
