+++ app/frontend/src/App.js
@@ -6,7 +6,6 @@ import Routes from './routes/Routes';
 import AuthenticationService from './services/AuthenticationService';
 import RouterRedirectTo from './components/simple/RouterRedirectTo';
 import {ROLE_CLIENT, ROLE_EMPLOYEE, ROLE_ADMIN, ACCESS_TOKEN, ACTIVE_ROLE} from './config/config';
-import Basket from './components/Basket';
 
 import ProfileSpeedDial from "./components/ProfileSpeedDial";
 import clsx from 'clsx';
